いろいろ細かい修正を入れた

+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: jtanaka <jtanaka@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/06 01:56:41 by jtanaka           #+#    #+#             */
-/*   Updated: 2020/10/16 01:25:11 by jtanaka          ###   ########.fr       */
+/*   Updated: 2020/10/22 17:51:18 by jtanaka          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,11 @@
 void	*ft_memchr(const void *s, int c, size_t n)
 {
 	unsigned char	*mem;
-	int				idx;
+	size_t			idx;
 
 	mem = (unsigned char*)s;
 	idx = 0;
-	while (idx < (int)n)
+	while (idx < n)
 	{
 		if (mem[idx] == (unsigned char)c)
 			return (mem + idx);
