+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: jtanaka <jtanaka@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/06 05:19:25 by jtanaka           #+#    #+#             */
-/*   Updated: 2020/10/11 04:46:13 by jtanaka          ###   ########.fr       */
+/*   Updated: 2020/10/22 18:10:57 by jtanaka          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 int		ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int				i;
+	size_t			i;
 	unsigned char	*str1;
 	unsigned char	*str2;
 
 	str1 = (unsigned char*)s1;
 	str2 = (unsigned char*)s2;
 	i = 0;
-	while (i < (int)n && (str1[i] || str2[i]))
+	while (i < n && (str1[i] || str2[i]))
 	{
 		if (str1[i] != str2[i])
 			return (str1[i] > str2[i] ? 1 : -1);
